Encapsulate creation of eda api clients
