+++ packages/libs/eda/src/lib/workspace/WorkspaceContainer.tsx
@@ -3,14 +3,15 @@ import { useRouteMatch } from 'react-router';
 
 import { TreeNode } from '@veupathdb/wdk-client/lib/Components/AttributeFilter/Types';
 
-import { EDAWorkspaceContainer, FieldWithMetadata } from '../core';
 import {
-  AnalysisClient,
-  ComputeClient,
-  DataClient,
-  DownloadClient,
-  SubsettingClient,
-} from '../core/api';
+  EDAWorkspaceContainer,
+  FieldWithMetadata,
+  useConfiguredAnalysisClient,
+  useConfiguredComputeClient,
+  useConfiguredDataClient,
+  useConfiguredDownloadClient,
+  useConfiguredSubsettingClient,
+} from '../core';
 import { VariableDescriptor } from '../core/types/variable';
 import { cx, findFirstVariable } from './Utils';
 
@@ -28,28 +29,25 @@ const useStyles = makeStyles({
 
 interface Props {
   studyId: string;
+  edaServiceUrl: string;
   analysisId?: string;
   children: ReactNode;
-  analysisClient: AnalysisClient;
-  computeClient: ComputeClient;
-  dataClient: DataClient;
-  downloadClient: DownloadClient;
-  subsettingClient: SubsettingClient;
   isStudyExplorerWorkspace?: boolean;
 }
 
 /** Allows a user to create a new analysis or edit an existing one. */
 export function WorkspaceContainer({
   studyId,
-  subsettingClient,
-  dataClient,
-  analysisClient,
-  downloadClient,
-  computeClient,
+  edaServiceUrl,
   children,
   isStudyExplorerWorkspace = false,
 }: Props) {
   const { url } = useRouteMatch();
+  const subsettingClient = useConfiguredSubsettingClient(edaServiceUrl);
+  const dataClient = useConfiguredDataClient(edaServiceUrl);
+  const analysisClient = useConfiguredAnalysisClient(edaServiceUrl);
+  const downloadClient = useConfiguredDownloadClient(edaServiceUrl);
+  const computeClient = useConfiguredComputeClient(edaServiceUrl);
 
   const initializeMakeVariableLink = useCallback(
     (fieldTree: TreeNode<FieldWithMetadata>) =>
